Copied the most recent py files to the top folder, created ZIP, and removed the awap subfolder
