--- awap/__init__.py
@@ -1,36 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-/***************************************************************************
- AwaP
-                                 A QGIS plugin
- This plugin takes in outlines of building blocks and the boundaray of area of interest, and calculates area-weighted average parameter (AwaP)
- Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
-                             -------------------
-        begin                : 2018-05-05
-        copyright            : (C) 2018 by Ivan Majic
-        email                : imajicos@gmail.com
-        git sha              : $Format:%H$
- ***************************************************************************/
-
-/***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
- This script initializes the plugin, making it known to QGIS.
-"""
-
-
-# noinspection PyPep8Naming
-def classFactory(iface):  # pylint: disable=invalid-name
-    """Load AwaP class from file AwaP.
-
-    :param iface: A QGIS interface instance.
-    :type iface: QgsInterface
-    """
-    #
-    from .awap import AwaP
-    return AwaP(iface)
