--- awap/awap_dialog.py
@@ -1,48 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-/***************************************************************************
- AwaPDialog
-                                 A QGIS plugin
- This plugin takes in outlines of building blocks and the boundaray of area of interest, and calculates area-weighted average parameter (AwaP)
- Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
-                             -------------------
-        begin                : 2018-05-05
-        git sha              : $Format:%H$
-        copyright            : (C) 2018 by Ivan Majic
-        email                : imajicos@gmail.com
- ***************************************************************************/
-
-/***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
-"""
-
-import os
-
-from PyQt5 import uic
-from PyQt5 import QtWidgets
-
-FORM_CLASS, _ = uic.loadUiType(os.path.join(
-    os.path.dirname(__file__), 'awap_dialog_base.ui'))
-
-
-class AwaPDialog(QtWidgets.QDialog, FORM_CLASS):
-    def __init__(self, parent=None):
-        """Constructor."""
-        super(AwaPDialog, self).__init__(parent)
-        # Set up the user interface from Designer.
-        # After setupUI you can access any designer object by doing
-        # self.<objectname>, and you can use autoconnect slots - see
-        # http://qt-project.org/doc/qt-4.8/designer-using-a-ui-file.html
-        # #widgets-and-dialogs-with-auto-connect
-        self.setupUi(self)
-
-        # this is my added part where I initially hide labels 6 and 7 (AwaP = value)
-        self.label_6.setVisible(False)
-        self.label_7.setVisible(False)
-        # self.label_9.setVisible(False)
