--- awap/resources.py
@@ -1,278 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Resource object code
-#
-# Created by: The Resource Compiler for PyQt5 (Qt v5.11.1)
-#
-# WARNING! All changes made in this file will be lost!
-
-from PyQt5 import QtCore
-
-qt_resource_data = b"\
-\x00\x00\x0d\x65\
-\x89\
-\x50\x4e\x47\x0d\x0a\x1a\x0a\x00\x00\x00\x0d\x49\x48\x44\x52\x00\
-\x00\x00\x17\x00\x00\x00\x18\x08\x00\x00\x00\x00\x34\x17\x39\xa9\
-\x00\x00\x00\x09\x70\x48\x59\x73\x00\x00\x2e\x23\x00\x00\x2e\x23\
-\x01\x78\xa5\x3f\x76\x00\x00\x0b\xc8\x69\x54\x58\x74\x58\x4d\x4c\
-\x3a\x63\x6f\x6d\x2e\x61\x64\x6f\x62\x65\x2e\x78\x6d\x70\x00\x00\
-\x00\x00\x00\x3c\x3f\x78\x70\x61\x63\x6b\x65\x74\x20\x62\x65\x67\
-\x69\x6e\x3d\x22\xef\xbb\xbf\x22\x20\x69\x64\x3d\x22\x57\x35\x4d\
-\x30\x4d\x70\x43\x65\x68\x69\x48\x7a\x72\x65\x53\x7a\x4e\x54\x63\
-\x7a\x6b\x63\x39\x64\x22\x3f\x3e\x20\x3c\x78\x3a\x78\x6d\x70\x6d\
-\x65\x74\x61\x20\x78\x6d\x6c\x6e\x73\x3a\x78\x3d\x22\x61\x64\x6f\
-\x62\x65\x3a\x6e\x73\x3a\x6d\x65\x74\x61\x2f\x22\x20\x78\x3a\x78\
-\x6d\x70\x74\x6b\x3d\x22\x41\x64\x6f\x62\x65\x20\x58\x4d\x50\x20\
-\x43\x6f\x72\x65\x20\x35\x2e\x36\x2d\x63\x31\x34\x35\x20\x37\x39\
-\x2e\x31\x36\x33\x34\x39\x39\x2c\x20\x32\x30\x31\x38\x2f\x30\x38\
-\x2f\x31\x33\x2d\x31\x36\x3a\x34\x30\x3a\x32\x32\x20\x20\x20\x20\
-\x20\x20\x20\x20\x22\x3e\x20\x3c\x72\x64\x66\x3a\x52\x44\x46\x20\
-\x78\x6d\x6c\x6e\x73\x3a\x72\x64\x66\x3d\x22\x68\x74\x74\x70\x3a\
-\x2f\x2f\x77\x77\x77\x2e\x77\x33\x2e\x6f\x72\x67\x2f\x31\x39\x39\
-\x39\x2f\x30\x32\x2f\x32\x32\x2d\x72\x64\x66\x2d\x73\x79\x6e\x74\
-\x61\x78\x2d\x6e\x73\x23\x22\x3e\x20\x3c\x72\x64\x66\x3a\x44\x65\
-\x73\x63\x72\x69\x70\x74\x69\x6f\x6e\x20\x72\x64\x66\x3a\x61\x62\
-\x6f\x75\x74\x3d\x22\x22\x20\x78\x6d\x6c\x6e\x73\x3a\x64\x63\x3d\
-\x22\x68\x74\x74\x70\x3a\x2f\x2f\x70\x75\x72\x6c\x2e\x6f\x72\x67\
-\x2f\x64\x63\x2f\x65\x6c\x65\x6d\x65\x6e\x74\x73\x2f\x31\x2e\x31\
-\x2f\x22\x20\x78\x6d\x6c\x6e\x73\x3a\x78\x6d\x70\x3d\x22\x68\x74\
-\x74\x70\x3a\x2f\x2f\x6e\x73\x2e\x61\x64\x6f\x62\x65\x2e\x63\x6f\
-\x6d\x2f\x78\x61\x70\x2f\x31\x2e\x30\x2f\x22\x20\x78\x6d\x6c\x6e\
-\x73\x3a\x78\x6d\x70\x4d\x4d\x3d\x22\x68\x74\x74\x70\x3a\x2f\x2f\
-\x6e\x73\x2e\x61\x64\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x78\x61\x70\
-\x2f\x31\x2e\x30\x2f\x6d\x6d\x2f\x22\x20\x78\x6d\x6c\x6e\x73\x3a\
-\x73\x74\x52\x65\x66\x3d\x22\x68\x74\x74\x70\x3a\x2f\x2f\x6e\x73\
-\x2e\x61\x64\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x78\x61\x70\x2f\x31\
-\x2e\x30\x2f\x73\x54\x79\x70\x65\x2f\x52\x65\x73\x6f\x75\x72\x63\
-\x65\x52\x65\x66\x23\x22\x20\x78\x6d\x6c\x6e\x73\x3a\x73\x74\x45\
-\x76\x74\x3d\x22\x68\x74\x74\x70\x3a\x2f\x2f\x6e\x73\x2e\x61\x64\
-\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x78\x61\x70\x2f\x31\x2e\x30\x2f\
-\x73\x54\x79\x70\x65\x2f\x52\x65\x73\x6f\x75\x72\x63\x65\x45\x76\
-\x65\x6e\x74\x23\x22\x20\x78\x6d\x6c\x6e\x73\x3a\x69\x6c\x6c\x75\
-\x73\x74\x72\x61\x74\x6f\x72\x3d\x22\x68\x74\x74\x70\x3a\x2f\x2f\
-\x6e\x73\x2e\x61\x64\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x69\x6c\x6c\
-\x75\x73\x74\x72\x61\x74\x6f\x72\x2f\x31\x2e\x30\x2f\x22\x20\x78\
-\x6d\x6c\x6e\x73\x3a\x70\x64\x66\x3d\x22\x68\x74\x74\x70\x3a\x2f\
-\x2f\x6e\x73\x2e\x61\x64\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x70\x64\
-\x66\x2f\x31\x2e\x33\x2f\x22\x20\x78\x6d\x6c\x6e\x73\x3a\x70\x68\
-\x6f\x74\x6f\x73\x68\x6f\x70\x3d\x22\x68\x74\x74\x70\x3a\x2f\x2f\
-\x6e\x73\x2e\x61\x64\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x70\x68\x6f\
-\x74\x6f\x73\x68\x6f\x70\x2f\x31\x2e\x30\x2f\x22\x20\x78\x6d\x6c\
-\x6e\x73\x3a\x74\x69\x66\x66\x3d\x22\x68\x74\x74\x70\x3a\x2f\x2f\
-\x6e\x73\x2e\x61\x64\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x74\x69\x66\
-\x66\x2f\x31\x2e\x30\x2f\x22\x20\x78\x6d\x6c\x6e\x73\x3a\x65\x78\
-\x69\x66\x3d\x22\x68\x74\x74\x70\x3a\x2f\x2f\x6e\x73\x2e\x61\x64\
-\x6f\x62\x65\x2e\x63\x6f\x6d\x2f\x65\x78\x69\x66\x2f\x31\x2e\x30\
-\x2f\x22\x20\x64\x63\x3a\x66\x6f\x72\x6d\x61\x74\x3d\x22\x69\x6d\
-\x61\x67\x65\x2f\x70\x6e\x67\x22\x20\x78\x6d\x70\x3a\x4d\x65\x74\
-\x61\x64\x61\x74\x61\x44\x61\x74\x65\x3d\x22\x32\x30\x31\x39\x2d\
-\x30\x31\x2d\x31\x38\x54\x31\x35\x3a\x30\x36\x3a\x33\x37\x2b\x31\
-\x31\x3a\x30\x30\x22\x20\x78\x6d\x70\x3a\x4d\x6f\x64\x69\x66\x79\
-\x44\x61\x74\x65\x3d\x22\x32\x30\x31\x39\x2d\x30\x31\x2d\x31\x38\
-\x54\x31\x35\x3a\x30\x36\x3a\x33\x37\x2b\x31\x31\x3a\x30\x30\x22\
-\x20\x78\x6d\x70\x3a\x43\x72\x65\x61\x74\x65\x44\x61\x74\x65\x3d\
-\x22\x32\x30\x31\x39\x2d\x30\x31\x2d\x31\x38\x54\x31\x35\x3a\x30\
-\x34\x3a\x35\x35\x2b\x31\x31\x3a\x30\x30\x22\x20\x78\x6d\x70\x3a\
-\x43\x72\x65\x61\x74\x6f\x72\x54\x6f\x6f\x6c\x3d\x22\x41\x64\x6f\
-\x62\x65\x20\x49\x6c\x6c\x75\x73\x74\x72\x61\x74\x6f\x72\x20\x43\
-\x43\x20\x32\x33\x2e\x30\x20\x28\x57\x69\x6e\x64\x6f\x77\x73\x29\
-\x22\x20\x78\x6d\x70\x4d\x4d\x3a\x49\x6e\x73\x74\x61\x6e\x63\x65\
-\x49\x44\x3d\x22\x78\x6d\x70\x2e\x69\x69\x64\x3a\x32\x65\x36\x32\
-\x34\x39\x36\x65\x2d\x37\x33\x65\x65\x2d\x38\x32\x34\x33\x2d\x61\
-\x31\x37\x33\x2d\x37\x37\x37\x38\x35\x61\x36\x35\x31\x37\x64\x34\
-\x22\x20\x78\x6d\x70\x4d\x4d\x3a\x44\x6f\x63\x75\x6d\x65\x6e\x74\
-\x49\x44\x3d\x22\x61\x64\x6f\x62\x65\x3a\x64\x6f\x63\x69\x64\x3a\
-\x70\x68\x6f\x74\x6f\x73\x68\x6f\x70\x3a\x38\x30\x62\x61\x65\x63\
-\x37\x61\x2d\x66\x35\x30\x65\x2d\x38\x34\x34\x39\x2d\x39\x36\x33\
-\x37\x2d\x33\x38\x66\x38\x61\x63\x31\x66\x65\x37\x65\x37\x22\x20\
-\x78\x6d\x70\x4d\x4d\x3a\x4f\x72\x69\x67\x69\x6e\x61\x6c\x44\x6f\
-\x63\x75\x6d\x65\x6e\x74\x49\x44\x3d\x22\x75\x75\x69\x64\x3a\x35\
-\x44\x32\x30\x38\x39\x32\x34\x39\x33\x42\x46\x44\x42\x31\x31\x39\
-\x31\x34\x41\x38\x35\x39\x30\x44\x33\x31\x35\x30\x38\x43\x38\x22\
-\x20\x78\x6d\x70\x4d\x4d\x3a\x52\x65\x6e\x64\x69\x74\x69\x6f\x6e\
-\x43\x6c\x61\x73\x73\x3d\x22\x70\x72\x6f\x6f\x66\x3a\x70\x64\x66\
-\x22\x20\x69\x6c\x6c\x75\x73\x74\x72\x61\x74\x6f\x72\x3a\x53\x74\
-\x61\x72\x74\x75\x70\x50\x72\x6f\x66\x69\x6c\x65\x3d\x22\x50\x72\
-\x69\x6e\x74\x22\x20\x70\x64\x66\x3a\x50\x72\x6f\x64\x75\x63\x65\
-\x72\x3d\x22\x41\x64\x6f\x62\x65\x20\x50\x44\x46\x20\x6c\x69\x62\
-\x72\x61\x72\x79\x20\x31\x30\x2e\x30\x31\x22\x20\x70\x68\x6f\x74\
-\x6f\x73\x68\x6f\x70\x3a\x43\x6f\x6c\x6f\x72\x4d\x6f\x64\x65\x3d\
-\x22\x31\x22\x20\x70\x68\x6f\x74\x6f\x73\x68\x6f\x70\x3a\x49\x43\
-\x43\x50\x72\x6f\x66\x69\x6c\x65\x3d\x22\x22\x20\x74\x69\x66\x66\
-\x3a\x49\x6d\x61\x67\x65\x57\x69\x64\x74\x68\x3d\x22\x31\x36\x35\
-\x22\x20\x74\x69\x66\x66\x3a\x49\x6d\x61\x67\x65\x4c\x65\x6e\x67\
-\x74\x68\x3d\x22\x31\x39\x34\x22\x20\x74\x69\x66\x66\x3a\x50\x68\
-\x6f\x74\x6f\x6d\x65\x74\x72\x69\x63\x49\x6e\x74\x65\x72\x70\x72\
-\x65\x74\x61\x74\x69\x6f\x6e\x3d\x22\x31\x22\x20\x74\x69\x66\x66\
-\x3a\x53\x61\x6d\x70\x6c\x65\x73\x50\x65\x72\x50\x69\x78\x65\x6c\
-\x3d\x22\x34\x22\x20\x74\x69\x66\x66\x3a\x58\x52\x65\x73\x6f\x6c\
-\x75\x74\x69\x6f\x6e\x3d\x22\x33\x30\x30\x2f\x31\x22\x20\x74\x69\
-\x66\x66\x3a\x59\x52\x65\x73\x6f\x6c\x75\x74\x69\x6f\x6e\x3d\x22\
-\x33\x30\x30\x2f\x31\x22\x20\x74\x69\x66\x66\x3a\x52\x65\x73\x6f\
-\x6c\x75\x74\x69\x6f\x6e\x55\x6e\x69\x74\x3d\x22\x32\x22\x20\x65\
-\x78\x69\x66\x3a\x45\x78\x69\x66\x56\x65\x72\x73\x69\x6f\x6e\x3d\
-\x22\x30\x32\x32\x31\x22\x20\x65\x78\x69\x66\x3a\x43\x6f\x6c\x6f\
-\x72\x53\x70\x61\x63\x65\x3d\x22\x36\x35\x35\x33\x35\x22\x20\x65\
-\x78\x69\x66\x3a\x50\x69\x78\x65\x6c\x58\x44\x69\x6d\x65\x6e\x73\
-\x69\x6f\x6e\x3d\x22\x31\x36\x35\x22\x20\x65\x78\x69\x66\x3a\x50\
-\x69\x78\x65\x6c\x59\x44\x69\x6d\x65\x6e\x73\x69\x6f\x6e\x3d\x22\
-\x31\x39\x34\x22\x3e\x20\x3c\x64\x63\x3a\x74\x69\x74\x6c\x65\x3e\
-\x20\x3c\x72\x64\x66\x3a\x41\x6c\x74\x3e\x20\x3c\x72\x64\x66\x3a\
-\x6c\x69\x20\x78\x6d\x6c\x3a\x6c\x61\x6e\x67\x3d\x22\x78\x2d\x64\
-\x65\x66\x61\x75\x6c\x74\x22\x3e\x50\x72\x69\x6e\x74\x3c\x2f\x72\
-\x64\x66\x3a\x6c\x69\x3e\x20\x3c\x2f\x72\x64\x66\x3a\x41\x6c\x74\
-\x3e\x20\x3c\x2f\x64\x63\x3a\x74\x69\x74\x6c\x65\x3e\x20\x3c\x78\
-\x6d\x70\x4d\x4d\x3a\x44\x65\x72\x69\x76\x65\x64\x46\x72\x6f\x6d\
-\x20\x73\x74\x52\x65\x66\x3a\x69\x6e\x73\x74\x61\x6e\x63\x65\x49\
-\x44\x3d\x22\x78\x6d\x70\x2e\x69\x69\x64\x3a\x31\x30\x38\x39\x34\
-\x63\x61\x64\x2d\x64\x34\x63\x36\x2d\x31\x38\x34\x63\x2d\x39\x33\
-\x34\x65\x2d\x64\x38\x37\x30\x65\x64\x33\x34\x35\x33\x35\x30\x22\
-\x20\x73\x74\x52\x65\x66\x3a\x64\x6f\x63\x75\x6d\x65\x6e\x74\x49\
-\x44\x3d\x22\x78\x6d\x70\x2e\x64\x69\x64\x3a\x37\x64\x32\x63\x36\
-\x63\x62\x61\x2d\x36\x63\x34\x33\x2d\x39\x37\x34\x31\x2d\x38\x30\
-\x65\x35\x2d\x62\x62\x61\x62\x35\x66\x63\x38\x30\x65\x36\x62\x22\
-\x20\x73\x74\x52\x65\x66\x3a\x6f\x72\x69\x67\x69\x6e\x61\x6c\x44\
-\x6f\x63\x75\x6d\x65\x6e\x74\x49\x44\x3d\x22\x75\x75\x69\x64\x3a\
-\x35\x44\x32\x30\x38\x39\x32\x34\x39\x33\x42\x46\x44\x42\x31\x31\
-\x39\x31\x34\x41\x38\x35\x39\x30\x44\x33\x31\x35\x30\x38\x43\x38\
-\x22\x20\x73\x74\x52\x65\x66\x3a\x72\x65\x6e\x64\x69\x74\x69\x6f\
-\x6e\x43\x6c\x61\x73\x73\x3d\x22\x70\x72\x6f\x6f\x66\x3a\x70\x64\
-\x66\x22\x2f\x3e\x20\x3c\x78\x6d\x70\x4d\x4d\x3a\x48\x69\x73\x74\
-\x6f\x72\x79\x3e\x20\x3c\x72\x64\x66\x3a\x53\x65\x71\x3e\x20\x3c\
-\x72\x64\x66\x3a\x6c\x69\x20\x73\x74\x45\x76\x74\x3a\x61\x63\x74\
-\x69\x6f\x6e\x3d\x22\x73\x61\x76\x65\x64\x22\x20\x73\x74\x45\x76\
-\x74\x3a\x69\x6e\x73\x74\x61\x6e\x63\x65\x49\x44\x3d\x22\x78\x6d\
-\x70\x2e\x69\x69\x64\x3a\x61\x35\x61\x35\x61\x32\x37\x38\x2d\x32\
-\x34\x30\x32\x2d\x64\x30\x34\x38\x2d\x38\x34\x31\x37\x2d\x37\x39\
-\x66\x66\x33\x35\x65\x32\x37\x64\x34\x65\x22\x20\x73\x74\x45\x76\
-\x74\x3a\x77\x68\x65\x6e\x3d\x22\x32\x30\x31\x39\x2d\x30\x31\x2d\
-\x31\x38\x54\x31\x34\x3a\x32\x39\x3a\x32\x39\x2b\x31\x31\x3a\x30\
-\x30\x22\x20\x73\x74\x45\x76\x74\x3a\x73\x6f\x66\x74\x77\x61\x72\
-\x65\x41\x67\x65\x6e\x74\x3d\x22\x41\x64\x6f\x62\x65\x20\x49\x6c\
-\x6c\x75\x73\x74\x72\x61\x74\x6f\x72\x20\x43\x43\x20\x32\x33\x2e\
-\x30\x20\x28\x57\x69\x6e\x64\x6f\x77\x73\x29\x22\x20\x73\x74\x45\
-\x76\x74\x3a\x63\x68\x61\x6e\x67\x65\x64\x3d\x22\x2f\x22\x2f\x3e\
-\x20\x3c\x72\x64\x66\x3a\x6c\x69\x20\x73\x74\x45\x76\x74\x3a\x61\
-\x63\x74\x69\x6f\x6e\x3d\x22\x73\x61\x76\x65\x64\x22\x20\x73\x74\
-\x45\x76\x74\x3a\x69\x6e\x73\x74\x61\x6e\x63\x65\x49\x44\x3d\x22\
-\x78\x6d\x70\x2e\x69\x69\x64\x3a\x31\x30\x38\x39\x34\x63\x61\x64\
-\x2d\x64\x34\x63\x36\x2d\x31\x38\x34\x63\x2d\x39\x33\x34\x65\x2d\
-\x64\x38\x37\x30\x65\x64\x33\x34\x35\x33\x35\x30\x22\x20\x73\x74\
-\x45\x76\x74\x3a\x77\x68\x65\x6e\x3d\x22\x32\x30\x31\x39\x2d\x30\
-\x31\x2d\x31\x38\x54\x31\x35\x3a\x30\x36\x3a\x33\x37\x2b\x31\x31\
-\x3a\x30\x30\x22\x20\x73\x74\x45\x76\x74\x3a\x73\x6f\x66\x74\x77\
-\x61\x72\x65\x41\x67\x65\x6e\x74\x3d\x22\x41\x64\x6f\x62\x65\x20\
-\x50\x68\x6f\x74\x6f\x73\x68\x6f\x70\x20\x43\x43\x20\x32\x30\x31\
-\x39\x20\x28\x57\x69\x6e\x64\x6f\x77\x73\x29\x22\x20\x73\x74\x45\
-\x76\x74\x3a\x63\x68\x61\x6e\x67\x65\x64\x3d\x22\x2f\x22\x2f\x3e\
-\x20\x3c\x72\x64\x66\x3a\x6c\x69\x20\x73\x74\x45\x76\x74\x3a\x61\
-\x63\x74\x69\x6f\x6e\x3d\x22\x63\x6f\x6e\x76\x65\x72\x74\x65\x64\
-\x22\x20\x73\x74\x45\x76\x74\x3a\x70\x61\x72\x61\x6d\x65\x74\x65\
-\x72\x73\x3d\x22\x66\x72\x6f\x6d\x20\x69\x6d\x61\x67\x65\x2f\x6a\
-\x70\x65\x67\x20\x74\x6f\x20\x69\x6d\x61\x67\x65\x2f\x70\x6e\x67\
-\x22\x2f\x3e\x20\x3c\x72\x64\x66\x3a\x6c\x69\x20\x73\x74\x45\x76\
-\x74\x3a\x61\x63\x74\x69\x6f\x6e\x3d\x22\x64\x65\x72\x69\x76\x65\
-\x64\x22\x20\x73\x74\x45\x76\x74\x3a\x70\x61\x72\x61\x6d\x65\x74\
-\x65\x72\x73\x3d\x22\x63\x6f\x6e\x76\x65\x72\x74\x65\x64\x20\x66\
-\x72\x6f\x6d\x20\x69\x6d\x61\x67\x65\x2f\x6a\x70\x65\x67\x20\x74\
-\x6f\x20\x69\x6d\x61\x67\x65\x2f\x70\x6e\x67\x22\x2f\x3e\x20\x3c\
-\x72\x64\x66\x3a\x6c\x69\x20\x73\x74\x45\x76\x74\x3a\x61\x63\x74\
-\x69\x6f\x6e\x3d\x22\x73\x61\x76\x65\x64\x22\x20\x73\x74\x45\x76\
-\x74\x3a\x69\x6e\x73\x74\x61\x6e\x63\x65\x49\x44\x3d\x22\x78\x6d\
-\x70\x2e\x69\x69\x64\x3a\x32\x65\x36\x32\x34\x39\x36\x65\x2d\x37\
-\x33\x65\x65\x2d\x38\x32\x34\x33\x2d\x61\x31\x37\x33\x2d\x37\x37\
-\x37\x38\x35\x61\x36\x35\x31\x37\x64\x34\x22\x20\x73\x74\x45\x76\
-\x74\x3a\x77\x68\x65\x6e\x3d\x22\x32\x30\x31\x39\x2d\x30\x31\x2d\
-\x31\x38\x54\x31\x35\x3a\x30\x36\x3a\x33\x37\x2b\x31\x31\x3a\x30\
-\x30\x22\x20\x73\x74\x45\x76\x74\x3a\x73\x6f\x66\x74\x77\x61\x72\
-\x65\x41\x67\x65\x6e\x74\x3d\x22\x41\x64\x6f\x62\x65\x20\x50\x68\
-\x6f\x74\x6f\x73\x68\x6f\x70\x20\x43\x43\x20\x32\x30\x31\x39\x20\
-\x28\x57\x69\x6e\x64\x6f\x77\x73\x29\x22\x20\x73\x74\x45\x76\x74\
-\x3a\x63\x68\x61\x6e\x67\x65\x64\x3d\x22\x2f\x22\x2f\x3e\x20\x3c\
-\x2f\x72\x64\x66\x3a\x53\x65\x71\x3e\x20\x3c\x2f\x78\x6d\x70\x4d\
-\x4d\x3a\x48\x69\x73\x74\x6f\x72\x79\x3e\x20\x3c\x74\x69\x66\x66\
-\x3a\x42\x69\x74\x73\x50\x65\x72\x53\x61\x6d\x70\x6c\x65\x3e\x20\
-\x3c\x72\x64\x66\x3a\x53\x65\x71\x3e\x20\x3c\x72\x64\x66\x3a\x6c\
-\x69\x3e\x38\x3c\x2f\x72\x64\x66\x3a\x6c\x69\x3e\x20\x3c\x72\x64\
-\x66\x3a\x6c\x69\x3e\x38\x3c\x2f\x72\x64\x66\x3a\x6c\x69\x3e\x20\
-\x3c\x72\x64\x66\x3a\x6c\x69\x3e\x38\x3c\x2f\x72\x64\x66\x3a\x6c\
-\x69\x3e\x20\x3c\x72\x64\x66\x3a\x6c\x69\x3e\x38\x3c\x2f\x72\x64\
-\x66\x3a\x6c\x69\x3e\x20\x3c\x2f\x72\x64\x66\x3a\x53\x65\x71\x3e\
-\x20\x3c\x2f\x74\x69\x66\x66\x3a\x42\x69\x74\x73\x50\x65\x72\x53\
-\x61\x6d\x70\x6c\x65\x3e\x20\x3c\x2f\x72\x64\x66\x3a\x44\x65\x73\
-\x63\x72\x69\x70\x74\x69\x6f\x6e\x3e\x20\x3c\x2f\x72\x64\x66\x3a\
-\x52\x44\x46\x3e\x20\x3c\x2f\x78\x3a\x78\x6d\x70\x6d\x65\x74\x61\
-\x3e\x20\x3c\x3f\x78\x70\x61\x63\x6b\x65\x74\x20\x65\x6e\x64\x3d\
-\x22\x72\x22\x3f\x3e\xcf\x07\x93\x46\x00\x00\x01\x43\x49\x44\x41\
-\x54\x28\x91\xad\x91\x3b\x2f\xc3\x71\x18\x85\x9f\xff\x5f\x93\xa6\
-\xd2\x10\xa4\x0b\x21\x06\x95\x60\xa8\xa4\x71\x59\x5c\x2a\x14\x83\
-\x7e\x00\x83\xd1\x28\x11\x1f\xc0\x27\x10\xb7\xc4\x60\x90\xd4\xc2\
-\x60\x68\x0c\x06\x86\x26\x8d\xc6\xb5\x83\xd4\xa0\x94\x18\x44\x04\
-\x49\x2b\x24\x44\xe9\x31\xfc\x94\xa5\x36\xef\xf2\x26\x4f\xf2\x9e\
-\x93\x73\x5e\x4b\x14\x1d\xbb\x38\xfe\x93\x3b\x00\xb8\x8e\x8d\x01\
-\x9c\xed\xb4\x64\x0e\xae\xca\xc6\x3b\x01\x49\xd2\xba\x59\x07\x03\
-\x10\x18\xa9\x62\x4a\x32\x60\x86\x3d\x49\x92\x08\x48\x0a\xb0\x2b\
-\xa3\x9f\x20\x6a\x64\x2b\x00\x66\x89\x18\xdf\xcb\x92\xba\xb0\xe1\
-\x79\x07\xe0\x21\x63\x78\xc4\x3b\x9d\x7a\x00\xc0\xf5\x01\x58\xb8\
-\x0d\x8f\x37\x06\x89\x01\xe0\xfc\x04\x6e\x08\x81\xa4\xb7\xf6\xac\
-\x18\x97\x24\xd5\x76\x48\xf2\x95\x4b\x36\x10\x75\x95\xd3\xb4\x66\
-\x0c\x1a\xf7\x27\xbb\x4f\x8e\x4c\xde\x78\x1b\x4c\x3c\xa7\x01\xb0\
-\x5c\x89\xca\x7b\xaf\xc9\x7b\x5c\xbf\xc9\x2d\x87\x0d\x00\x29\x5f\
-\xac\xd0\xc3\xfb\x79\x7e\x8e\x52\x56\x47\x01\xdb\x59\x28\x48\x5a\
-\x6e\x93\xa4\x2e\x72\x92\xaa\x7b\x4d\x6e\xd9\xb0\xd1\x0a\x30\x48\
-\x12\x78\x75\xfc\xf6\x7c\x37\x0a\x10\x62\x1b\xc8\xe5\x7f\x74\x2e\
-\xbc\xe6\xd2\x53\x23\xc9\x1e\xfa\xd6\xe1\xb4\x99\x85\x17\xe9\x35\
-\x0c\xf3\xe9\x45\x58\x79\x94\x24\x95\xf4\x67\x87\x93\x7d\x6e\x9e\
-\x96\x82\x3d\x4f\xee\xad\x90\x3f\xe9\xf7\x00\x58\xff\xf4\xf7\x2f\
-\x5d\x55\x9f\x8f\x79\xfb\x29\xa4\x00\x00\x00\x00\x49\x45\x4e\x44\
-\xae\x42\x60\x82\
-"
-
-qt_resource_name = b"\
-\x00\x07\
-\x07\x3b\xe0\xb3\
-\x00\x70\
-\x00\x6c\x00\x75\x00\x67\x00\x69\x00\x6e\x00\x73\
-\x00\x04\
-\x00\x06\x8d\x80\
-\x00\x61\
-\x00\x77\x00\x61\x00\x70\
-\x00\x08\
-\x0a\x61\x5a\xa7\
-\x00\x69\
-\x00\x63\x00\x6f\x00\x6e\x00\x2e\x00\x70\x00\x6e\x00\x67\
-"
-
-qt_resource_struct_v1 = b"\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
-\x00\x00\x00\x14\x00\x02\x00\x00\x00\x01\x00\x00\x00\x03\
-\x00\x00\x00\x22\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-"
-
-qt_resource_struct_v2 = b"\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x01\
-\x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
-\x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x14\x00\x02\x00\x00\x00\x01\x00\x00\x00\x03\
-\x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x22\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01\x68\x5f\x54\xa8\x62\
-"
-
-qt_version = [int(v) for v in QtCore.qVersion().split('.')]
-if qt_version < [5, 8, 0]:
-    rcc_version = 1
-    qt_resource_struct = qt_resource_struct_v1
-else:
-    rcc_version = 2
-    qt_resource_struct = qt_resource_struct_v2
-
-def qInitResources():
-    QtCore.qRegisterResourceData(rcc_version, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-def qCleanupResources():
-    QtCore.qUnregisterResourceData(rcc_version, qt_resource_struct, qt_resource_name, qt_resource_data)
-
-qInitResources()
