+++ awap_worker.py
